Move config type out of Vars.py
